Change how alerts work, improve security

+++ admin/modify.php
@@ -1,19 +1,22 @@
 <?php
 
   session_start();
+  $messageType = "success";
   $id = $_POST["id"];
   $table = $_POST["table"];
   $type = $_POST["type"];
+  $title = $type;
   $url = $_SERVER['HTTP_REFERER'];
+  include("assets/php/actions.php");
 
-  if (isset($_POST["new"])) {
+  if (isset($_POST["new"]) && in_array("create", $actions)) {
       $sql = "INSERT INTO yourmood.${table} () VALUES ()";
       $message = "Successfully created ${type}";
-  } elseif (isset($_POST["delete"])) {
+  } elseif (isset($_POST["delete"]) && in_array("delete", $actions)) {
       $sql = "DELETE FROM yourmood.{$table} WHERE id='{$id}'";
       $message = "Successfully deleted ${type}";
       $url = "{$type}s?type=${table}";
-  } elseif (isset($_POST["edit"])) {
+  } elseif (isset($_POST["edit"]) && in_array("edit", $actions)) {
       $statement = "";
       foreach ($_POST["values"] as $key => $val) {
           $statement .= $key . " = '" . addslashes($val) . "',";
@@ -21,10 +24,14 @@
       $statement = substr($statement, 0, -1);
       $sql = "UPDATE yourmood.${table} SET {$statement} WHERE id = '{$id}'";
       $message = "Successfully updated ${type}";
-  } elseif (isset($_POST["deleteall"])) {
+      $url = in_array("view", $actions) ? "view/?type=${table}&title=${type}&id=${id}" : "{$type}s?type=${table}";
+  } elseif (isset($_POST["deleteall"]) && in_array("deleteall", $actions)) {
     $sql = "DELETE FROM yourmood.{$table}";
     $message = "Successfully deleted all ${type}s";
     $url = "{$type}s?type=${table}";
+  } else {
+    $messageType = "danger";
+    $message = "An error occured";
   }
 
 
@@ -34,5 +41,5 @@
   }
   $conn->close();
 
-  $_SESSION["alert"] = $message;
+  $_SESSION[$messageType] = $message;
   header("location: {$url}");
